split transposer from layout_autotune.h
