+++ paddle/phi/kernels/autotune/layout_autotune.h
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 #pragma once
+#include <memory>
 #include <set>
-#include "paddle/fluid/imperative/var_helper.h"
-#include "paddle/phi/backends/gpu/gpu_info.h"
-#include "paddle/phi/core/enforce.h"
-#include "paddle/phi/core/errors.h"
-#include "paddle/phi/kernels/autotune/layout_agnostic_ops.h"
+#include "paddle/fluid/framework/op_info.h"
+#include "paddle/phi/common/layout.h"
+#include "paddle/phi/common/place.h"
+#include "paddle/phi/core/compat/type_defs.h"
 
 namespace phi {
 namespace autotune {
@@ -26,38 +26,6 @@ namespace autotune {
 class DenseTensor;
 
 using DataLayout = paddle::experimental::DataLayout;
-
-template <typename VarType>
-std::shared_ptr<VarType> TraceTransposeOp(
-    const std::shared_ptr<VarType>& var,
-    const DataLayout layout,
-    const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-  std::vector<int> axis;
-  if (layout == DataLayout::NHWC) {
-    axis = {0, 2, 3, 1};
-  } else {
-    axis = {0, 3, 1, 2};
-  }
-  paddle::imperative::NameVarMap<VarType> ins = {{"X", {var}}};
-  auto out =
-      std::shared_ptr<VarType>(new VarType(tracer->GenerateUniqueName()));
-  auto x_shape =
-      std::shared_ptr<VarType>(new VarType(tracer->GenerateUniqueName()));
-  paddle::imperative::NameVarMap<VarType> outs = {{"Out", {out}},
-                                                  {"XShape", {x_shape}}};
-  paddle::framework::AttributeMap attrs = {{"axis", axis}};
-  tracer->TraceOp("transpose2", ins, outs, std::move(attrs));
-  paddle::imperative::SetDataLayout(out, layout);
-  VLOG(4) << "Transpose " << paddle::imperative::GetNameFromVar(var) << "["
-          << paddle::framework::DataLayoutToString(
-                 paddle::imperative::GetDataLayout(var))
-          << "]"
-          << " to " << paddle::imperative::GetNameFromVar(out) << "["
-          << paddle::framework::DataLayoutToString(
-                 paddle::imperative::GetDataLayout(out))
-          << "]";
-  return out;
-}
 
 class LayoutAutoTune {
  public:
@@ -86,15 +54,59 @@ class LayoutAutoTune {
 
  private:
   LayoutAutoTune() {
-    layout_agnostic_ops_ =
-        paddle::imperative::LayoutAutotuneOperators::Instance()
-            .GetLayoutAgnosticOps();
-    heavily_layout_sensitive_ops_ =
-        paddle::imperative::LayoutAutotuneOperators::Instance()
-            .GetHeavilyLayoutSensitiveOps();
-    lightly_layout_sensitive_ops_ =
-        paddle::imperative::LayoutAutotuneOperators::Instance()
-            .GetLightlyLayoutSensitiveOps();
+    const auto& op_info = paddle::framework::OpInfoMap::Instance().map();
+    for (auto it = op_info.begin(); it != op_info.end(); it++) {
+      // only record forwrd operators
+      if (it->first.find("_grad") != std::string::npos) {
+        continue;
+      }
+
+      // some normalization operators such as instance_norm and layer_norm
+      // do not have data_format attr, but are layout sensitive.
+      if (it->first.find("norm") != std::string::npos) {
+        layout_agnostic_ops_.emplace(it->first);
+        continue;
+      }
+
+      auto* attr_checker = it->second.Checker();
+      if (attr_checker) {
+        auto attrs = attr_checker->GetDefaultAttrMap();
+        if (attrs.find("data_format") != attrs.end() ||
+            attrs.find("data_layout") != attrs.end()) {
+          VLOG(4) << "Heavily layout sensitive OP: " << it->first;
+          heavily_layout_sensitive_ops_.emplace(it->first);
+          continue;
+        }
+
+        // Attribute name is fuzzy matched, such as start and start_axis.
+        bool layout_agnostic = true;
+        for (auto& attr : attrs) {
+          auto attr_name = attr.first;
+          VLOG(6) << "OP: " << it->first << " Attr Name: " << attr_name;
+          if (attr_name.find("axis") != std::string::npos ||
+              attr_name.find("axes") != std::string::npos ||
+              attr_name.find("dim") != std::string::npos ||
+              attr_name.find("start") != std::string::npos ||
+              attr_name.find("end") != std::string::npos) {
+            VLOG(4) << "Lightly layout sensitive OP: " << it->first;
+            layout_agnostic = false;
+            lightly_layout_sensitive_ops_.emplace(it->first);
+            break;
+          }
+        }
+
+        if (layout_agnostic) {
+          VLOG(4) << "Layout agnostic_ops: " << it->first;
+          layout_agnostic_ops_.emplace(it->first);
+        }
+      }
+    }
+
+    VLOG(3) << "The number of layout agnostic OPs: "
+            << layout_agnostic_ops_.size() << ", heavily layout sensitive OPs: "
+            << heavily_layout_sensitive_ops_.size()
+            << ", lightly layout sensitive OPs: "
+            << lightly_layout_sensitive_ops_.size();
   }
 
   bool use_layout_autotune_{false};
@@ -107,352 +119,6 @@ class LayoutAutoTune {
 
   DataLayout layout_{DataLayout::UNDEFINED};
 };
-
-template <typename VarType>
-class LayoutTransposer {
- public:
-  explicit LayoutTransposer(const std::string& type) : type_(type) {}
-
-  virtual ~LayoutTransposer() {}
-
-  LayoutTransposer(const LayoutTransposer&) = delete;
-  LayoutTransposer& operator=(const LayoutTransposer&) = delete;
-
-  virtual paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    VLOG(3) << "Optimze Layout agnostic op: " << type_;
-    auto in_layout = DataLayout::UNDEFINED;
-    for (auto& pair : ins) {
-      for (auto& var : pair.second) {
-        // Once the any input is desired layout, we set in_layout is desired
-        // layout.
-        if (paddle::imperative::GetDataLayout(var) ==
-            LayoutAutoTune::Instance().GetDesiredLayout()) {
-          in_layout = LayoutAutoTune::Instance().GetDesiredLayout();
-          break;
-        }
-      }
-    }
-    SetVarsLayout(outs, in_layout);
-    return ins;
-  }
-
-  // Set inputs, outputs and attributes to be optimized for the transposer.
-  // Those may respectively be a subset of the corresponding original argument
-  // of the operator.
-  void SetArguments(const std::vector<std::string>& ins,
-                    const std::vector<std::string>& outs,
-                    const std::vector<std::string>& attrs) {
-    ins_ = ins;
-    outs_ = outs;
-    attrs_ = attrs;
-  }
-
-  // Set the variables's layout to the specified layout.
-  // If outs_ is not specified, it means all outputs of the operator
-  // will be considered. Otherwise, it only set layout for the specified output.
-  void SetVarsLayout(const paddle::imperative::NameVarMap<VarType>& outs,
-                     DataLayout layout) const {
-    if (outs_.empty()) {
-      for (auto& pair : outs) {
-        for (auto& var : pair.second) {
-          paddle::imperative::SetDataLayout(var, layout);
-        }
-      }
-    } else {
-      for (auto& name : outs_) {
-        auto out_vars = outs.at(name);
-        for (auto& var : out_vars) {
-          paddle::imperative::SetDataLayout(var, layout);
-        }
-      }
-    }
-  }
-
-  const std::vector<std::string>& Inputs() const { return ins_; }
-  const std::vector<std::string>& Outputs() const { return outs_; }
-  const std::vector<std::string>& Attributes() const { return attrs_; }
-
-  const std::string& Type() { return type_; }
-
- protected:
-  std::string type_{};
-  std::vector<std::string> ins_{};
-  std::vector<std::string> outs_{};
-  std::vector<std::string> attrs_{};
-};
-
-template <typename VarType>
-class ElementwiseOpTransposer : public LayoutTransposer<VarType> {
- public:
-  explicit ElementwiseOpTransposer(const std::string& type)
-      : LayoutTransposer<VarType>(type) {}
-
-  paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    // [Why we need the this?]
-    // The Elementwise Ops has a axis attr, it is to support broadcast.
-    // When bias_attr of Conv is not false, the elementwise_add will be
-    // appended, and the axis will be set to the channel dimension.
-
-    // If the axis is set to the channel dimension, the attr transformation
-    // is necessary. Otherwise, it will fall back to the LayoutTransposer::Run.
-    auto desired_layout = LayoutAutoTune::Instance().GetDesiredLayout();
-    if (attrs->find("axis") != attrs->end() &&
-        boost::get<int>((*attrs)["axis"]) != -1) {
-      VLOG(3) << "Optimze layout agnostic op " << this->Type();
-      if (desired_layout == DataLayout::NHWC) {
-        (*attrs)["axis"] = 3;
-      } else if (desired_layout == DataLayout::NCHW) {
-        (*attrs)["axis"] = 1;
-      } else {
-        PADDLE_ENFORCE_EQ(
-            desired_layout,
-            DataLayout::UNDEFINED,
-            phi::errors::PreconditionNotMet("DataLayout is unsupport."));
-      }
-      this->SetVarsLayout(outs, desired_layout);
-      return ins;
-    } else {
-      return LayoutTransposer<VarType>::Run(ins, outs, attrs, tracer);
-    }
-  }
-};
-
-/*
- * Both functionality and performance are affected by data layout.
- * Such as operators with data_format attribute.
- */
-template <typename VarType>
-class HeavilyLayoutSensitiveOpTransposer : public LayoutTransposer<VarType> {
- public:
-  explicit HeavilyLayoutSensitiveOpTransposer(const std::string& type)
-      : LayoutTransposer<VarType>(type) {}
-
-  paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    VLOG(3) << "Optimze heavily layout sensitive op " << this->Type();
-    paddle::imperative::NameVarMap<VarType> new_ins(ins);
-
-    // Step 1: Adjust the data_layout attr to the desired layout
-    auto desired_layout = LayoutAutoTune::Instance().GetDesiredLayout();
-    std::string desired_layout_str = paddle::framework::DataLayoutToString(
-        LayoutAutoTune::Instance().GetDesiredLayout());
-    if (attrs->find("data_format") != attrs->end() &&
-        boost::get<std::string>((*attrs)["data_format"]) !=
-            desired_layout_str) {
-      VLOG(4) << "Origin layout attr: "
-              << boost::get<std::string>((*attrs)["data_format"])
-              << ", Desired layout attr: " << desired_layout_str;
-      (*attrs)["data_format"] = desired_layout_str;
-    } else if (attrs->find("data_layout") != attrs->end() &&
-               boost::get<std::string>((*attrs)["data_layout"]) !=
-                   desired_layout_str) {
-      VLOG(4) << "Origin layout attr: "
-              << boost::get<std::string>((*attrs)["data_layout"])
-              << ", Desired layout attr: " << desired_layout_str;
-      (*attrs)["data_layout"] = desired_layout_str;
-    }
-
-    // Step 2: Transpose the specified input for Op and set the transposed var's
-    // layout.
-    for (auto& name : this->Inputs()) {
-      auto& in_vars = new_ins[name];
-      for (auto& var : in_vars) {
-        auto var_layout = paddle::imperative::GetDataLayout(var);
-        if (var_layout != desired_layout) {
-          var = TraceTransposeOp(var, DataLayout::NHWC, tracer);
-        }
-      }
-    }
-
-    // Step 3: Set the Op's layout sensitive outs var.
-    this->SetVarsLayout(outs, desired_layout);
-
-    return new_ins;
-  }
-};
-
-/*
- * The functionality may be affected layout transformation before them.
- * Such as operators with axis attribute.
- */
-template <typename VarType>
-class LightlyLayoutSensitiveOpTransposer : public LayoutTransposer<VarType> {
- public:
-  explicit LightlyLayoutSensitiveOpTransposer(const std::string& type)
-      : LayoutTransposer<VarType>(type) {}
-
-  paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    VLOG(3) << "Optimze lightly layout sensitive op " << this->Type();
-    paddle::imperative::NameVarMap<VarType> new_ins(ins);
-    // If input's layout is not tuned, transformation is unnecessary.
-    // If input's layout is already tuned, it will be transformed back to NCHW.
-    // TODO(zhangting): The op of this type should be adapted to the previous
-    // operator output data layout. Currently only a few operators are
-    // supported, and transposers need to be carefully designed to ensure that
-    // they do not cause exceptions.
-    for (auto& pair : new_ins) {
-      for (auto& var : pair.second) {
-        auto var_layout = paddle::imperative::GetDataLayout(var);
-        if (var_layout == LayoutAutoTune::Instance().GetDesiredLayout()) {
-          var = TraceTransposeOp(var, DataLayout::NCHW, tracer);
-        }
-      }
-    }
-    return new_ins;
-  }
-};
-
-template <typename VarType>
-class TransposeOpTransposer
-    : public LightlyLayoutSensitiveOpTransposer<VarType> {
- public:
-  explicit TransposeOpTransposer(const std::string& type)
-      : LightlyLayoutSensitiveOpTransposer<VarType>(type) {}
-
-  paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    VLOG(3) << "Optimze lightly layout sensitive op " << this->Type();
-    // When the input layout is the desired format, it means that there
-    // is a transpose layer in the network, it is better to transpose
-    // the result to the original format.
-    // Instead of actually inserting a transpose Op, we fuse the inserted
-    // transpose Op with the current transpose Op by transforming 'axis' attr.
-    auto& in_var = ins.at("X")[0];
-    auto var_layout = paddle::imperative::GetDataLayout(in_var);
-    if (var_layout == LayoutAutoTune::Instance().GetDesiredLayout()) {
-      auto axis = boost::get<std::vector<int>>((*attrs)["axis"]);
-      // NHWC->NCHW, permutaion will be set as follows.
-      std::vector<int> perm = {0, 3, 1, 2};
-      // fuse the transpose Op by transforming axis.
-      if (axis != perm) {
-        std::vector<int> fusion_axis = {
-            perm[axis[0]], perm[axis[1]], perm[axis[2]], perm[axis[3]]};
-        (*attrs)["axis"] = fusion_axis;
-      }
-    }
-    return ins;
-  }
-};
-
-template <typename VarType>
-class FlattenOpTransposer : public LightlyLayoutSensitiveOpTransposer<VarType> {
- public:
-  explicit FlattenOpTransposer(const std::string& type)
-      : LightlyLayoutSensitiveOpTransposer<VarType>(type) {}
-
-  paddle::imperative::NameVarMap<VarType> Run(
-      const paddle::imperative::NameVarMap<VarType>& ins,
-      const paddle::imperative::NameVarMap<VarType>& outs,
-      paddle::framework::AttributeMap* attrs,
-      const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-    VLOG(3) << "Optimze lightly layout sensitive op " << this->Type();
-    // Flatten the C, H, W dimensions will not affect functionality.
-    // So transformation is unnecessary. But in other cases, it needs to
-    // fall back to the LightlyLayoutSensitiveOpTransposer.
-    auto start_axis = boost::get<int>((*attrs)["start_axis"]);
-    auto stop_axis = boost::get<int>((*attrs)["stop_axis"]);
-    if (paddle::imperative::GetDataLayout(ins.at("X")[0]) ==
-            LayoutAutoTune::Instance().GetDesiredLayout() &&
-        start_axis == 1 && stop_axis == 3) {
-      return ins;
-    } else {
-      return LightlyLayoutSensitiveOpTransposer<VarType>::Run(
-          ins, outs, attrs, tracer);
-    }
-  }
-};
-
-template <typename VarType>
-paddle::imperative::NameVarMap<VarType> LayoutOptimizer(
-    const std::string& op_type,
-    const paddle::imperative::NameVarMap<VarType>& ins,
-    const paddle::imperative::NameVarMap<VarType>& outs,
-    paddle::framework::AttributeMap* attrs,
-    const paddle::platform::Place& place,
-    const std::shared_ptr<paddle::imperative::Tracer>& tracer) {
-  if (!LayoutAutoTune::Instance().UseLayoutAutoTune()) {
-    return ins;
-  }
-
-  // When layout autotuning is enabled, the tuner will check the desired layout.
-  // (1) If the desired layout is undefined, and there is no convolutional
-  // layers, layout optimization is unnecessary. Otherwise, the desired layout
-  // will be set to the best layout only when these is a convolutional layer
-  // with
-  // NCHW-Layout and the TensorCore is available.
-  // (2) If the desired layout is defined, run the transposer.
-
-  if (LayoutAutoTune::Instance().GetDesiredLayout() == DataLayout::UNDEFINED) {
-    // Layout autotune only supports model with convolutional layers
-    if (op_type != "conv2d") {
-      return ins;
-    } else {
-      if (boost::get<std::string>((*attrs)["data_format"]) == "NCHW" &&
-          paddle::platform::is_gpu_place(place) &&
-          phi::backends::gpu::TensorCoreAvailable()) {
-        LayoutAutoTune::Instance().SetDesiredLayout(DataLayout::NHWC);
-        VLOG(3) << "Tune the layout from "
-                << boost::get<std::string>((*attrs)["data_format"]) << " to "
-                << paddle::framework::DataLayoutToString(
-                       LayoutAutoTune::Instance().GetDesiredLayout());
-      } else {
-        LayoutAutoTune::Instance().DisableLayoutAutoTune();
-        return ins;
-      }
-    }
-  }
-
-  std::shared_ptr<LayoutTransposer<VarType>> transposer = nullptr;
-  if (op_type == "conv2d") {
-    transposer =
-        std::make_shared<HeavilyLayoutSensitiveOpTransposer<VarType>>(op_type);
-    transposer->SetArguments({"Input"}, {"Output"}, {"data_format"});
-  } else if (op_type == "batch_norm") {
-    transposer =
-        std::make_shared<HeavilyLayoutSensitiveOpTransposer<VarType>>(op_type);
-    transposer->SetArguments({"X"}, {"Y"}, {"data_layout"});
-  } else if (op_type == "pool2d") {
-    transposer =
-        std::make_shared<HeavilyLayoutSensitiveOpTransposer<VarType>>(op_type);
-    transposer->SetArguments({"X"}, {"Out"}, {"data_format"});
-  } else if (op_type == "transpose2") {
-    transposer = std::make_shared<TransposeOpTransposer<VarType>>(op_type);
-  } else if (op_type == "flatten_contiguous_range") {
-    transposer = std::make_shared<FlattenOpTransposer<VarType>>(op_type);
-  } else if (op_type.find("elementwise_") != std::string::npos) {
-    transposer = std::make_shared<ElementwiseOpTransposer<VarType>>(op_type);
-  } else if (LayoutAutoTune::Instance().IsLayoutAgnostic(op_type)) {
-    transposer = std::make_shared<LayoutTransposer<VarType>>(op_type);
-  } else if (LayoutAutoTune::Instance().IsLightlyLayoutSensitive(op_type)) {
-    transposer =
-        std::make_shared<LightlyLayoutSensitiveOpTransposer<VarType>>(op_type);
-  } else {
-    PADDLE_ENFORCE_NOT_NULL(
-        transposer,
-        phi::errors::Unimplemented("%s 's LayoutTransposer is unimplemented.",
-                                   op_type));
-  }
-
-  return transposer->Run(ins, outs, attrs, tracer);
-}
 
 }  // namespace autotune
 }  // namespace phi
